+++ bookkeeping/src/sdk/sdkconfiguration.py
@@ -26,9 +26,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '4.1.0'
-    gen_version: str = '2.277.0'
-    user_agent: str = 'speakeasy-sdk/python 4.1.0 2.277.0 1.0.0 wingspan_bookkeeping'
+    sdk_version: str = '4.2.0'
+    gen_version: str = '2.279.1'
+    user_agent: str = 'speakeasy-sdk/python 4.2.0 2.279.1 1.0.0 wingspan_bookkeeping'
     retry_config: RetryConfig = None
     _hooks: SDKHooks = None