+++ bookkeeping/RELEASES.md
@@ -120,3 +120,13 @@ Based on:
 - [python v4.1.0] bookkeeping
 ### Releases
 - [PyPI v4.1.0] https://pypi.org/project/wingspan_bookkeeping/4.1.0 - bookkeeping
+
+## 2024-03-11 01:27:03
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.205.0 (2.279.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0] bookkeeping
+### Releases
+- [PyPI v4.2.0] https://pypi.org/project/wingspan_bookkeeping/4.2.0 - bookkeeping